refactor(defaulting webhook): breaking the original interface instead of introducing new interface
