+++ pkg/webhook/admission/validator_warn_test.go
@@ -1,3 +1,19 @@
+/*
+Copyright 2021 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
 package admission
 
 import (
@@ -17,23 +33,24 @@ import (
 	"k8s.io/client-go/kubernetes/scheme"
 )
 
-var fakeValidatorWarnVK = schema.GroupVersionKind{Group: "foo.test.org", Version: "v1", Kind: "fakeValidatorWarn"}
+var fakeValidatorVK = schema.GroupVersionKind{Group: "foo.test.org", Version: "v1", Kind: "fakeValidator"}
 
-var _ = Describe("validatingWarnHandler", func() {
+var _ = Describe("validatingHandler", func() {
 
-	decoder, _ := NewDecoder(scheme.Scheme)
+	decoder := NewDecoder(scheme.Scheme)
 
 	Context("when dealing with successful results without warning", func() {
-		f := &admissiontest.FakeValidatorWarn{ErrorToReturn: nil, GVKToReturn: fakeValidatorWarnVK, WarningsToReturn: nil}
-		handler := validatingWarnHandler{validatorWarn: f, decoder: decoder}
+		f := &admissiontest.FakeValidator{ErrorToReturn: nil, GVKToReturn: fakeValidatorVK, WarningsToReturn: nil}
+		handler := validatingHandler{validator: f, decoder: decoder}
 
 		It("should return 200 in response when create succeeds", func() {
+
 			response := handler.Handle(context.TODO(), Request{
 				AdmissionRequest: admissionv1.AdmissionRequest{
 					Operation: admissionv1.Create,
 					Object: runtime.RawExtension{
 						Raw:    []byte("{}"),
-						Object: handler.validatorWarn,
+						Object: handler.validator,
 					},
 				},
 			})
@@ -49,11 +66,11 @@ var _ = Describe("validatingWarnHandler", func() {
 					Operation: admissionv1.Update,
 					Object: runtime.RawExtension{
 						Raw:    []byte("{}"),
-						Object: handler.validatorWarn,
+						Object: handler.validator,
 					},
 					OldObject: runtime.RawExtension{
 						Raw:    []byte("{}"),
-						Object: handler.validatorWarn,
+						Object: handler.validator,
 					},
 				},
 			})
@@ -68,7 +85,7 @@ var _ = Describe("validatingWarnHandler", func() {
 					Operation: admissionv1.Delete,
 					OldObject: runtime.RawExtension{
 						Raw:    []byte("{}"),
-						Object: handler.validatorWarn,
+						Object: handler.validator,
 					},
 				},
 			})
@@ -80,11 +97,11 @@ var _ = Describe("validatingWarnHandler", func() {
 	const warningMessage = "warning message"
 	const anotherWarningMessage = "another warning message"
 	Context("when dealing with successful results with warning", func() {
-		f := &admissiontest.FakeValidatorWarn{ErrorToReturn: nil, GVKToReturn: fakeValidatorWarnVK, WarningsToReturn: []string{
+		f := &admissiontest.FakeValidator{ErrorToReturn: nil, GVKToReturn: fakeValidatorVK, WarningsToReturn: []string{
 			warningMessage,
 			anotherWarningMessage,
 		}}
-		handler := validatingWarnHandler{validatorWarn: f, decoder: decoder}
+		handler := validatingHandler{validator: f, decoder: decoder}
 
 		It("should return 200 in response when create succeeds, with warning messages", func() {
 			response := handler.Handle(context.TODO(), Request{
@@ -92,7 +109,7 @@ var _ = Describe("validatingWarnHandler", func() {
 					Operation: admissionv1.Create,
 					Object: runtime.RawExtension{
 						Raw:    []byte("{}"),
-						Object: handler.validatorWarn,
+						Object: handler.validator,
 					},
 				},
 			})
@@ -110,11 +127,11 @@ var _ = Describe("validatingWarnHandler", func() {
 					Operation: admissionv1.Update,
 					Object: runtime.RawExtension{
 						Raw:    []byte("{}"),
-						Object: handler.validatorWarn,
+						Object: handler.validator,
 					},
 					OldObject: runtime.RawExtension{
 						Raw:    []byte("{}"),
-						Object: handler.validatorWarn,
+						Object: handler.validator,
 					},
 				},
 			})
@@ -131,7 +148,7 @@ var _ = Describe("validatingWarnHandler", func() {
 					Operation: admissionv1.Delete,
 					OldObject: runtime.RawExtension{
 						Raw:    []byte("{}"),
-						Object: handler.validatorWarn,
+						Object: handler.validator,
 					},
 				},
 			})
@@ -142,16 +159,16 @@ var _ = Describe("validatingWarnHandler", func() {
 		})
 	})
 
-	Context("when dealing with Status errors", func() {
-
+	Context("when dealing with Status errors, with warning messages", func() {
+		// Status error would overwrite the warning messages, so no warning messages should be observed.
 		expectedError := &apierrors.StatusError{
 			ErrStatus: metav1.Status{
 				Message: "some message",
 				Code:    http.StatusUnprocessableEntity,
 			},
 		}
-		f := &admissiontest.FakeValidatorWarn{ErrorToReturn: expectedError, GVKToReturn: fakeValidatorVK, WarningsToReturn: nil}
-		handler := validatingWarnHandler{validatorWarn: f, decoder: decoder}
+		f := &admissiontest.FakeValidator{ErrorToReturn: expectedError, GVKToReturn: fakeValidatorVK, WarningsToReturn: []string{warningMessage, anotherWarningMessage}}
+		handler := validatingHandler{validator: f, decoder: decoder}
 
 		It("should propagate the Status from ValidateCreate's return value to the HTTP response", func() {
 
@@ -160,7 +177,81 @@ var _ = Describe("validatingWarnHandler", func() {
 					Operation: admissionv1.Create,
 					Object: runtime.RawExtension{
 						Raw:    []byte("{}"),
-						Object: handler.validatorWarn,
+						Object: handler.validator,
+					},
+				},
+			})
+
+			Expect(response.Allowed).Should(BeFalse())
+			Expect(response.Result.Code).Should(Equal(expectedError.Status().Code))
+			Expect(*response.Result).Should(Equal(expectedError.Status()))
+			Expect(response.AdmissionResponse.Warnings).Should(BeEmpty())
+
+		})
+
+		It("should propagate the Status from ValidateUpdate's return value to the HTTP response", func() {
+
+			response := handler.Handle(context.TODO(), Request{
+				AdmissionRequest: admissionv1.AdmissionRequest{
+					Operation: admissionv1.Update,
+					Object: runtime.RawExtension{
+						Raw:    []byte("{}"),
+						Object: handler.validator,
+					},
+					OldObject: runtime.RawExtension{
+						Raw:    []byte("{}"),
+						Object: handler.validator,
+					},
+				},
+			})
+
+			Expect(response.Allowed).Should(BeFalse())
+			Expect(response.Result.Code).Should(Equal(expectedError.Status().Code))
+			Expect(*response.Result).Should(Equal(expectedError.Status()))
+			Expect(response.AdmissionResponse.Warnings).Should(BeEmpty())
+
+		})
+
+		It("should propagate the Status from ValidateDelete's return value to the HTTP response", func() {
+
+			response := handler.Handle(context.TODO(), Request{
+				AdmissionRequest: admissionv1.AdmissionRequest{
+					Operation: admissionv1.Delete,
+					OldObject: runtime.RawExtension{
+						Raw:    []byte("{}"),
+						Object: handler.validator,
+					},
+				},
+			})
+
+			Expect(response.Allowed).Should(BeFalse())
+			Expect(response.Result.Code).Should(Equal(expectedError.Status().Code))
+			Expect(*response.Result).Should(Equal(expectedError.Status()))
+			Expect(response.AdmissionResponse.Warnings).Should(BeEmpty())
+
+		})
+
+	})
+
+	Context("when dealing with Status errors, without warning messages", func() {
+
+		expectedError := &apierrors.StatusError{
+			ErrStatus: metav1.Status{
+				Message: "some message",
+				Code:    http.StatusUnprocessableEntity,
+			},
+		}
+		f := &admissiontest.FakeValidator{ErrorToReturn: expectedError, GVKToReturn: fakeValidatorVK, WarningsToReturn: nil}
+		handler := validatingHandler{validator: f, decoder: decoder}
+
+		It("should propagate the Status from ValidateCreate's return value to the HTTP response", func() {
+
+			response := handler.Handle(context.TODO(), Request{
+				AdmissionRequest: admissionv1.AdmissionRequest{
+					Operation: admissionv1.Create,
+					Object: runtime.RawExtension{
+						Raw:    []byte("{}"),
+						Object: handler.validator,
 					},
 				},
 			})
@@ -178,11 +269,11 @@ var _ = Describe("validatingWarnHandler", func() {
 					Operation: admissionv1.Update,
 					Object: runtime.RawExtension{
 						Raw:    []byte("{}"),
-						Object: handler.validatorWarn,
+						Object: handler.validator,
 					},
 					OldObject: runtime.RawExtension{
 						Raw:    []byte("{}"),
-						Object: handler.validatorWarn,
+						Object: handler.validator,
 					},
 				},
 			})
@@ -200,7 +291,7 @@ var _ = Describe("validatingWarnHandler", func() {
 					Operation: admissionv1.Delete,
 					OldObject: runtime.RawExtension{
 						Raw:    []byte("{}"),
-						Object: handler.validatorWarn,
+						Object: handler.validator,
 					},
 				},
 			})
@@ -216,8 +307,8 @@ var _ = Describe("validatingWarnHandler", func() {
 	Context("when dealing with non-status errors, without warning messages", func() {
 
 		expectedError := errors.New("some error")
-		f := &admissiontest.FakeValidatorWarn{ErrorToReturn: expectedError, GVKToReturn: fakeValidatorVK}
-		handler := validatingWarnHandler{validatorWarn: f, decoder: decoder}
+		f := &admissiontest.FakeValidator{ErrorToReturn: expectedError, GVKToReturn: fakeValidatorVK}
+		handler := validatingHandler{validator: f, decoder: decoder}
 
 		It("should return 403 response when ValidateCreate with error message embedded", func() {
 
@@ -226,13 +317,14 @@ var _ = Describe("validatingWarnHandler", func() {
 					Operation: admissionv1.Create,
 					Object: runtime.RawExtension{
 						Raw:    []byte("{}"),
-						Object: handler.validatorWarn,
+						Object: handler.validator,
 					},
 				},
 			})
 			Expect(response.Allowed).Should(BeFalse())
 			Expect(response.Result.Code).Should(Equal(int32(http.StatusForbidden)))
-			Expect(string(response.Result.Reason)).Should(Equal(expectedError.Error()))
+			Expect(response.Result.Reason).Should(Equal(metav1.StatusReasonForbidden))
+			Expect(response.Result.Message).Should(Equal(expectedError.Error()))
 
 		})
 
@@ -243,17 +335,18 @@ var _ = Describe("validatingWarnHandler", func() {
 					Operation: admissionv1.Update,
 					Object: runtime.RawExtension{
 						Raw:    []byte("{}"),
-						Object: handler.validatorWarn,
+						Object: handler.validator,
 					},
 					OldObject: runtime.RawExtension{
 						Raw:    []byte("{}"),
-						Object: handler.validatorWarn,
+						Object: handler.validator,
 					},
 				},
 			})
 			Expect(response.Allowed).Should(BeFalse())
 			Expect(response.Result.Code).Should(Equal(int32(http.StatusForbidden)))
-			Expect(string(response.Result.Reason)).Should(Equal(expectedError.Error()))
+			Expect(response.Result.Reason).Should(Equal(metav1.StatusReasonForbidden))
+			Expect(response.Result.Message).Should(Equal(expectedError.Error()))
 
 		})
 
@@ -263,21 +356,22 @@ var _ = Describe("validatingWarnHandler", func() {
 					Operation: admissionv1.Delete,
 					OldObject: runtime.RawExtension{
 						Raw:    []byte("{}"),
-						Object: handler.validatorWarn,
+						Object: handler.validator,
 					},
 				},
 			})
 			Expect(response.Allowed).Should(BeFalse())
 			Expect(response.Result.Code).Should(Equal(int32(http.StatusForbidden)))
-			Expect(string(response.Result.Reason)).Should(Equal(expectedError.Error()))
+			Expect(response.Result.Reason).Should(Equal(metav1.StatusReasonForbidden))
+			Expect(response.Result.Message).Should(Equal(expectedError.Error()))
 		})
 	})
 
 	Context("when dealing with non-status errors, with warning messages", func() {
 
 		expectedError := errors.New("some error")
-		f := &admissiontest.FakeValidatorWarn{ErrorToReturn: expectedError, GVKToReturn: fakeValidatorVK, WarningsToReturn: []string{warningMessage, anotherWarningMessage}}
-		handler := validatingWarnHandler{validatorWarn: f, decoder: decoder}
+		f := &admissiontest.FakeValidator{ErrorToReturn: expectedError, GVKToReturn: fakeValidatorVK, WarningsToReturn: []string{warningMessage, anotherWarningMessage}}
+		handler := validatingHandler{validator: f, decoder: decoder}
 
 		It("should return 403 response when ValidateCreate with error message embedded", func() {
 
@@ -286,13 +380,14 @@ var _ = Describe("validatingWarnHandler", func() {
 					Operation: admissionv1.Create,
 					Object: runtime.RawExtension{
 						Raw:    []byte("{}"),
-						Object: handler.validatorWarn,
+						Object: handler.validator,
 					},
 				},
 			})
 			Expect(response.Allowed).Should(BeFalse())
 			Expect(response.Result.Code).Should(Equal(int32(http.StatusForbidden)))
-			Expect(string(response.Result.Reason)).Should(Equal(expectedError.Error()))
+			Expect(response.Result.Reason).Should(Equal(metav1.StatusReasonForbidden))
+			Expect(response.Result.Message).Should(Equal(expectedError.Error()))
 			Expect(response.AdmissionResponse.Warnings).Should(ContainElement(warningMessage))
 			Expect(response.AdmissionResponse.Warnings).Should(ContainElement(anotherWarningMessage))
 		})
@@ -304,17 +399,18 @@ var _ = Describe("validatingWarnHandler", func() {
 					Operation: admissionv1.Update,
 					Object: runtime.RawExtension{
 						Raw:    []byte("{}"),
-						Object: handler.validatorWarn,
+						Object: handler.validator,
 					},
 					OldObject: runtime.RawExtension{
 						Raw:    []byte("{}"),
-						Object: handler.validatorWarn,
+						Object: handler.validator,
 					},
 				},
 			})
 			Expect(response.Allowed).Should(BeFalse())
 			Expect(response.Result.Code).Should(Equal(int32(http.StatusForbidden)))
-			Expect(string(response.Result.Reason)).Should(Equal(expectedError.Error()))
+			Expect(response.Result.Reason).Should(Equal(metav1.StatusReasonForbidden))
+			Expect(response.Result.Message).Should(Equal(expectedError.Error()))
 			Expect(response.AdmissionResponse.Warnings).Should(ContainElement(warningMessage))
 			Expect(response.AdmissionResponse.Warnings).Should(ContainElement(anotherWarningMessage))
 
@@ -326,17 +422,26 @@ var _ = Describe("validatingWarnHandler", func() {
 					Operation: admissionv1.Delete,
 					OldObject: runtime.RawExtension{
 						Raw:    []byte("{}"),
-						Object: handler.validatorWarn,
+						Object: handler.validator,
 					},
 				},
 			})
 			Expect(response.Allowed).Should(BeFalse())
 			Expect(response.Result.Code).Should(Equal(int32(http.StatusForbidden)))
-			Expect(string(response.Result.Reason)).Should(Equal(expectedError.Error()))
+			Expect(response.Result.Reason).Should(Equal(metav1.StatusReasonForbidden))
+			Expect(response.Result.Message).Should(Equal(expectedError.Error()))
 			Expect(response.AdmissionResponse.Warnings).Should(ContainElement(warningMessage))
 			Expect(response.AdmissionResponse.Warnings).Should(ContainElement(anotherWarningMessage))
 
 		})
 	})
+
+	PIt("should return 400 in response when create fails on decode", func() {})
+
+	PIt("should return 400 in response when update fails on decoding new object", func() {})
+
+	PIt("should return 400 in response when update fails on decoding old object", func() {})
+
+	PIt("should return 400 in response when delete fails on decode", func() {})
 
 })
